requirements.txt
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0
rembg==2.0.50
numpy==1.26.3
